+++ pywren_ibm_cloud/invoker.py
@@ -25,7 +25,7 @@ from pywren_ibm_cloud.utils import version_str
 from pywren_ibm_cloud.version import __version__
 from concurrent.futures import ThreadPoolExecutor
 from pywren_ibm_cloud.config import extract_storage_config, extract_compute_config
-from pywren_ibm_cloud.future import ResponseFuture, CallState
+from pywren_ibm_cloud.future import ResponseFuture
 from pywren_ibm_cloud.storage.utils import create_output_key, create_status_key
 
 logger = logging.getLogger(__name__)
@@ -47,7 +47,7 @@ class FunctionInvoker:
         self.compute_handlers = []
         cb = self.compute_config['backend']
         regions = self.compute_config[cb].get('region')
-        if type(regions) == list:
+        if regions and type(regions) == list:
             for region in regions:
                 compute_config = self.compute_config.copy()
                 compute_config[cb]['region'] = region
@@ -128,31 +128,26 @@ class FunctionInvoker:
 
         ########################
 
-        def invoke(executor_id, job_id, call_id, func_key, invoke_metadata, data_key, data_byte_range):
+        def invoke(executor_id, job_id, call_id, func_key, job_metadata, data_key, data_byte_range):
 
             output_key = create_output_key(self.storage_config['prefix'], executor_id, job_id, call_id)
             status_key = create_status_key(self.storage_config['prefix'], executor_id, job_id, call_id)
 
-            payload = {
-                'config': self.pywren_config,
-                'log_level': self.log_level,
-                'func_key': func_key,
-                'data_key': data_key,
-                'output_key': output_key,
-                'status_key': status_key,
-                'execution_timeout': job.execution_timeout,
-                'data_byte_range': data_byte_range,
-                'executor_id': executor_id,
-                'job_id': job_id,
-                'call_id': call_id,
-                'pywren_version': __version__}
+            payload = {'config': self.pywren_config,
+                       'log_level': self.log_level,
+                       'func_key': func_key,
+                       'data_key': data_key,
+                       'output_key': output_key,
+                       'status_key': status_key,
+                       'extra_env': job.extra_env,
+                       'execution_timeout': job.execution_timeout,
+                       'data_byte_range': data_byte_range,
+                       'executor_id': executor_id,
+                       'job_id': job_id,
+                       'call_id': call_id,
+                       'host_submit_time': time.time(),
+                       'pywren_version': __version__}
 
-            if job.extra_env is not None:
-                logger.debug("Extra environment vars {}".format(job.extra_env))
-                payload['extra_env'] = job.extra_env
-
-            host_submit_time = time.time()
-            payload['host_submit_time'] = host_submit_time
             # do the invocation
             compute_handler = random.choice(self.compute_handlers)
             activation_id = compute_handler.invoke(job.runtime_name, job.runtime_memory, payload)
@@ -161,14 +156,9 @@ class FunctionInvoker:
                 raise Exception("ExecutorID {} | JobID {} - Retrying mechanism finished with no success. "
                                 "Failed to invoke the job".format(executor_id, job_id))
 
-            invoke_metadata['activation_id'] = activation_id
-            invoke_metadata['invoke_time'] = time.time() - host_submit_time
-
-            invoke_metadata.update(payload)
-            del invoke_metadata['config']
-
-            fut = ResponseFuture(call_id, job_id, executor_id, activation_id, self.storage_config, invoke_metadata)
-            fut._set_state(CallState.invoked)
+            job_metadata['activation_id'] = activation_id
+            fut = ResponseFuture(executor_id, job_id, call_id, self.storage_config, job_metadata)
+            fut._set_state(ResponseFuture.State.Invoked)
 
             return fut
 
@@ -181,7 +171,7 @@ class FunctionInvoker:
                 data_byte_range = job.data_ranges[i]
                 future = executor.submit(invoke, self.executor_id,
                                          job.job_id, call_id, job.func_key,
-                                         job.host_job_meta.copy(),
+                                         job.metadata.copy(),
                                          job.data_key, data_byte_range)
                 call_futures.append(future)