+++ pywren_ibm_cloud/job/__init__.py
@@ -1,2 +1,3 @@
 from pywren_ibm_cloud.job.job import create_map_job
 from pywren_ibm_cloud.job.job import create_reduce_job
+from pywren_ibm_cloud.job.job import JobState